Mini Project 2/requirements.txt
Brotli==1.0.9
click==7.1.2
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
paho-mqtt==1.5.1
pandas==1.1.5
plotly==4.14.3
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
six==1.15.0
Werkzeug==1.0.1
